src/requirements.txt
flask
psycopg2-binary
flask-socketio
psycopg2-binary
Flask-SocketIO
python-engineio
python-socketio
pytest
requests